requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.3
django-cors-headers==3.1.1
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
drf-generators==0.4.0
gunicorn==19.7.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
mysqlclient==1.4.4
openapi-codec==1.3.2
protobuf==3.10.0
PyJWT==1.6.1
pytz==2018.3
requests==2.22.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.6
